fix: copilot comments, UI appearance, bugs
