+++ cellacdc/fonts.py
@@ -1,0 +1,7 @@
+from qtpy.QtGui import QFont
+
+font = QFont()
+font.setPixelSize(12)
+italicFont = QFont()
+italicFont.setPixelSize(12)
+italicFont.setItalic(True)
